requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
ebooklib==0.18
PyYAML==6.0.2
scrapy==2.11.2
shub==2.15.1
